+++ src/util/__init__.py
@@ -1,6 +1,3 @@
-from .sha256 import sha256_hash
-from .validator import is_valid_phone, is_valid_student_id, is_valid_year, is_valid_semester
-from .get_scsc_global_status import get_scsc_global_status
-from .create_uuid import create_uuid
-from .get_file_extension import get_file_extension
-from .user_role import get_user_role_level
+from .validator import sha256_hash, create_uuid, is_valid_phone, is_valid_student_id, is_valid_year, is_valid_semester
+from .helper import get_file_extension
+from .get_from_db import get_user_role_level, get_scsc_global_status
