--- src/util/create_uuid.py
@@ -1,5 +1,0 @@
-import uuid
-
-
-def create_uuid() -> str:
-    return str(uuid.uuid4())
